print.c: Fix format string for Windows
dwMajorVersion and dwMinorVersion are DWORDS. Cast them to `unsigned
long` to avoid a format error.

TEST=Builds on Cygwin (Windows 10, amd64, gcc 11.3.0)

Reviewed-on: https://review.coreboot.org/c/flashrom/+/66547
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ print.c
@@ -524,7 +524,8 @@ static void print_sysinfo(void)
 	/* Tell Windows which version of the structure we want. */
 	osvi.dwOSVersionInfoSize = sizeof(OSVERSIONINFOEX);
 	if (GetVersionEx((OSVERSIONINFO*) &osvi))
-		msg_ginfo(" %lu.%lu", osvi.dwMajorVersion, osvi.dwMinorVersion);
+		msg_ginfo(" %lu.%lu", (unsigned long)osvi.dwMajorVersion,
+			(unsigned long)osvi.dwMinorVersion);
 	else
 		msg_ginfo(" unknown version");
 	GetSystemInfo(&si);
